packages/rag_service/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0
Pillow>=10.0.0
fastapi>=0.100.0
pydantic>=2.0.0

